move event and service check code to their own file, cleanup
