+++ statsd/statsd.go
@@ -27,7 +27,6 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
-	"io"
 	"math/rand"
 	"os"
 	"strconv"
@@ -221,7 +220,7 @@ func (c *Client) format(name string, value interface{}, suffix []byte, tags []st
 		buf = strconv.AppendFloat(buf, rate, 'f', -1, 64)
 	}
 
-	buf = appendTagString(buf, c.Tags, tags)
+	buf = appendTags(buf, c.Tags, tags)
 
 	// non-zeroing copy to avoid referencing a larger than necessary underlying array
 	return append([]byte(nil), buf...)
@@ -480,274 +479,4 @@ func (c *Client) Close() error {
 	}
 
 	return c.writer.Close()
-}
-
-// Events support
-// EventAlertType and EventAlertPriority became exported types after this issue was submitted: https://github.com/DataDog/datadog-go/issues/41
-// The reason why they got exported is so that client code can directly use the types.
-
-// EventAlertType is the alert type for events
-type EventAlertType string
-
-const (
-	// Info is the "info" AlertType for events
-	Info EventAlertType = "info"
-	// Error is the "error" AlertType for events
-	Error EventAlertType = "error"
-	// Warning is the "warning" AlertType for events
-	Warning EventAlertType = "warning"
-	// Success is the "success" AlertType for events
-	Success EventAlertType = "success"
-)
-
-// EventPriority is the event priority for events
-type EventPriority string
-
-const (
-	// Normal is the "normal" Priority for events
-	Normal EventPriority = "normal"
-	// Low is the "low" Priority for events
-	Low EventPriority = "low"
-)
-
-// An Event is an object that can be posted to your DataDog event stream.
-type Event struct {
-	// Title of the event.  Required.
-	Title string
-	// Text is the description of the event.  Required.
-	Text string
-	// Timestamp is a timestamp for the event.  If not provided, the dogstatsd
-	// server will set this to the current time.
-	Timestamp time.Time
-	// Hostname for the event.
-	Hostname string
-	// AggregationKey groups this event with others of the same key.
-	AggregationKey string
-	// Priority of the event.  Can be statsd.Low or statsd.Normal.
-	Priority EventPriority
-	// SourceTypeName is a source type for the event.
-	SourceTypeName string
-	// AlertType can be statsd.Info, statsd.Error, statsd.Warning, or statsd.Success.
-	// If absent, the default value applied by the dogstatsd server is Info.
-	AlertType EventAlertType
-	// Tags for the event.
-	Tags []string
-}
-
-// NewEvent creates a new event with the given title and text.  Error checking
-// against these values is done at send-time, or upon running e.Check.
-func NewEvent(title, text string) *Event {
-	return &Event{
-		Title: title,
-		Text:  text,
-	}
-}
-
-// Check verifies that an event is valid.
-func (e Event) Check() error {
-	if len(e.Title) == 0 {
-		return fmt.Errorf("statsd.Event title is required")
-	}
-	if len(e.Text) == 0 {
-		return fmt.Errorf("statsd.Event text is required")
-	}
-	return nil
-}
-
-// Encode returns the dogstatsd wire protocol representation for an event.
-// Tags may be passed which will be added to the encoded output but not to
-// the Event's list of tags, eg. for default tags.
-func (e Event) Encode(tags ...string) (string, error) {
-	err := e.Check()
-	if err != nil {
-		return "", err
-	}
-	text := e.escapedText()
-
-	var buffer bytes.Buffer
-	buffer.WriteString("_e{")
-	buffer.WriteString(strconv.FormatInt(int64(len(e.Title)), 10))
-	buffer.WriteRune(',')
-	buffer.WriteString(strconv.FormatInt(int64(len(text)), 10))
-	buffer.WriteString("}:")
-	buffer.WriteString(e.Title)
-	buffer.WriteRune('|')
-	buffer.WriteString(text)
-
-	if !e.Timestamp.IsZero() {
-		buffer.WriteString("|d:")
-		buffer.WriteString(strconv.FormatInt(int64(e.Timestamp.Unix()), 10))
-	}
-
-	if len(e.Hostname) != 0 {
-		buffer.WriteString("|h:")
-		buffer.WriteString(e.Hostname)
-	}
-
-	if len(e.AggregationKey) != 0 {
-		buffer.WriteString("|k:")
-		buffer.WriteString(e.AggregationKey)
-
-	}
-
-	if len(e.Priority) != 0 {
-		buffer.WriteString("|p:")
-		buffer.WriteString(string(e.Priority))
-	}
-
-	if len(e.SourceTypeName) != 0 {
-		buffer.WriteString("|s:")
-		buffer.WriteString(e.SourceTypeName)
-	}
-
-	if len(e.AlertType) != 0 {
-		buffer.WriteString("|t:")
-		buffer.WriteString(string(e.AlertType))
-	}
-
-	writeTagString(&buffer, tags, e.Tags)
-
-	return buffer.String(), nil
-}
-
-// ServiceCheckStatus support
-type ServiceCheckStatus byte
-
-const (
-	// Ok is the "ok" ServiceCheck status
-	Ok ServiceCheckStatus = 0
-	// Warn is the "warning" ServiceCheck status
-	Warn ServiceCheckStatus = 1
-	// Critical is the "critical" ServiceCheck status
-	Critical ServiceCheckStatus = 2
-	// Unknown is the "unknown" ServiceCheck status
-	Unknown ServiceCheckStatus = 3
-)
-
-// An ServiceCheck is an object that contains status of DataDog service check.
-type ServiceCheck struct {
-	// Name of the service check.  Required.
-	Name string
-	// Status of service check.  Required.
-	Status ServiceCheckStatus
-	// Timestamp is a timestamp for the serviceCheck.  If not provided, the dogstatsd
-	// server will set this to the current time.
-	Timestamp time.Time
-	// Hostname for the serviceCheck.
-	Hostname string
-	// A message describing the current state of the serviceCheck.
-	Message string
-	// Tags for the serviceCheck.
-	Tags []string
-}
-
-// NewServiceCheck creates a new serviceCheck with the given name and status.  Error checking
-// against these values is done at send-time, or upon running sc.Check.
-func NewServiceCheck(name string, status ServiceCheckStatus) *ServiceCheck {
-	return &ServiceCheck{
-		Name:   name,
-		Status: status,
-	}
-}
-
-// Check verifies that an event is valid.
-func (sc ServiceCheck) Check() error {
-	if len(sc.Name) == 0 {
-		return fmt.Errorf("statsd.ServiceCheck name is required")
-	}
-	if byte(sc.Status) < 0 || byte(sc.Status) > 3 {
-		return fmt.Errorf("statsd.ServiceCheck status has invalid value")
-	}
-	return nil
-}
-
-// Encode returns the dogstatsd wire protocol representation for an serviceCheck.
-// Tags may be passed which will be added to the encoded output but not to
-// the Event's list of tags, eg. for default tags.
-func (sc ServiceCheck) Encode(tags ...string) (string, error) {
-	err := sc.Check()
-	if err != nil {
-		return "", err
-	}
-	message := sc.escapedMessage()
-
-	var buffer bytes.Buffer
-	buffer.WriteString("_sc|")
-	buffer.WriteString(sc.Name)
-	buffer.WriteRune('|')
-	buffer.WriteString(strconv.FormatInt(int64(sc.Status), 10))
-
-	if !sc.Timestamp.IsZero() {
-		buffer.WriteString("|d:")
-		buffer.WriteString(strconv.FormatInt(int64(sc.Timestamp.Unix()), 10))
-	}
-
-	if len(sc.Hostname) != 0 {
-		buffer.WriteString("|h:")
-		buffer.WriteString(sc.Hostname)
-	}
-
-	writeTagString(&buffer, tags, sc.Tags)
-
-	if len(message) != 0 {
-		buffer.WriteString("|m:")
-		buffer.WriteString(message)
-	}
-
-	return buffer.String(), nil
-}
-
-func (e Event) escapedText() string {
-	return strings.Replace(e.Text, "\n", "\\n", -1)
-}
-
-func (sc ServiceCheck) escapedMessage() string {
-	msg := strings.Replace(sc.Message, "\n", "\\n", -1)
-	return strings.Replace(msg, "m:", `m\:`, -1)
-}
-
-func removeNewlines(str string) string {
-	return strings.Replace(str, "\n", "", -1)
-}
-
-func writeTagString(w io.Writer, tagList1, tagList2 []string) {
-	// the tag lists may be shared with other callers, so we cannot modify
-	// them in any way (which means we cannot append to them either)
-	// therefore we must make an entirely separate copy just for this call
-	totalLen := len(tagList1) + len(tagList2)
-	if totalLen == 0 {
-		return
-	}
-	tags := make([]string, 0, totalLen)
-	tags = append(tags, tagList1...)
-	tags = append(tags, tagList2...)
-
-	io.WriteString(w, "|#")
-	io.WriteString(w, removeNewlines(tags[0]))
-	for _, tag := range tags[1:] {
-		io.WriteString(w, ",")
-		io.WriteString(w, removeNewlines(tag))
-	}
-}
-
-func appendTagString(buf []byte, tagList1, tagList2 []string) []byte {
-	if len(tagList1) == 0 {
-		if len(tagList2) == 0 {
-			return buf
-		}
-		tagList1 = tagList2
-		tagList2 = nil
-	}
-
-	buf = append(buf, "|#"...)
-	buf = appendWithoutNewlines(buf, tagList1[0])
-	for _, tag := range tagList1[1:] {
-		buf = append(buf, ',')
-		buf = appendWithoutNewlines(buf, tag)
-	}
-	for _, tag := range tagList2 {
-		buf = append(buf, ',')
-		buf = appendWithoutNewlines(buf, tag)
-	}
-	return buf
 }
